refactor: 移动 PermissionsMixin 位置
